requirements.txt
bitsandbytes
peft
trl
accelerate
datasets
transformers
Flask
Flask-Cors
python-dotenv
google-generativeai
requests